Addressing comments on commit c23e288
I also moved the seed dispersal struct to ST_seedDispersal.h.
